# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
  description: 'TaskY API — canonical contract generated from springdoc on a dev-profile boot (Observed). Regenerate: boot
    api with dev profile, GET /api-docs, convert to YAML.'
  x-generated: 2026-09-26, feat OIDC code exchange + timer e2e
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
...
  /api/v1/activity-templates/{templateId}/use:
    post:
      tags:
      - activity-template-controller
      operationId: use
      parameters:
      - name: templateId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UseActivityTemplateRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/GeneratedActivityResponse'
...
  /api/v1/activities/{activityId}/templates:
    post:
      tags:
      - activity-template-controller
      operationId: create_9
      parameters:
      - name: activityId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateActivityTemplateRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ActivityTemplateResponse'
...
  /api/v1/projects/{projectId}/activity-templates:
    get:
      tags:
      - activity-template-controller
      operationId: list_9
      parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ActivityTemplateResponse'
...
components:
  schemas:
...
    UseActivityTemplateRequest:
      type: object
      properties:
        occurrenceAt:
          type: string
          format: date-time
      required:
      - occurrenceAt
    GeneratedActivityResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        projectId:
          type: string
          format: uuid
        startDatetime:
          type: string
          format: date-time
        endDatetime:
          type: string
          format: date-time
...
    CreateActivityTemplateRequest:
      type: object
      properties:
        templateId:
          type: string
          format: uuid
        name:
          type: string
          minLength: 1
        recurrence:
          $ref: '#/components/schemas/RecurrenceRequest'
    RecurrenceRequest:
      type: object
      properties:
        frequency:
          type: string
          enum:
          - DAILY
          - WEEKLY
        interval:
          type: integer
          format: int32
          minimum: 1
        timezone:
          type: string
          minLength: 1
        nextOccurrence:
          type: string
          format: date-time
      required:
      - frequency
      - nextOccurrence
    ActivityTemplateResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        projectId:
          type: string
          format: uuid
        name:
          type: string
        version:
          type: integer
          format: int32
        title:
          type: string
        description:
          type: string
        weight:
          type: integer
          format: int32
        durationSeconds:
          type: integer
          format: int64
        estimatedSeconds:
          type: integer
          format: int64
        assignedToMembershipId:
          type: string
          format: uuid
        recurrence:
          $ref: '#/components/schemas/RecurrenceResponse'
        createdAt:
          type: string
          format: date-time
    RecurrenceResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        frequency:
          type: string
          enum:
          - DAILY
          - WEEKLY
        interval:
          type: integer
          format: int32
        timezone:
          type: string
        nextOccurrence:
          type: string
          format: date-time
        active:
          type: boolean
...